fix compilation for amd64

+++ plugins/HLTMonBTagMuSource.cc
@@ -2,8 +2,8 @@
  *
  *  DQM source for BJet HLT paths
  *
- *  $Date: 2009/09/10 13:07:02 $
- *  $Revision: 1.7 $
+ *  $Date: 2009/10/02 08:43:08 $
+ *  $Revision: 1.8 $
  *  \author Andrea Bocci, Pisa
  *
  */
@@ -235,7 +235,7 @@ void HLTMonBTagMuSource::analyze(const edm::Event & event, const edm::EventSetup
     m_plotRates->Fill( 5. );    // HLT accepted
 
   if ((latest > m_L1FilterIndex) and h_L2Jets.isValid()) {
-    unsigned int size = std::min(h_L2Jets->size(), m_size);
+    unsigned int size = std::min((unsigned int) h_L2Jets->size(), m_size);
     for (unsigned int i = 0; i < size; ++i) {
       const reco::Jet & jet = (*h_L2Jets)[i];
       m_plotL2JetsEnergy->Fill( jet.energy() );
@@ -247,7 +247,7 @@ void HLTMonBTagMuSource::analyze(const edm::Event & event, const edm::EventSetup
     }
   }
   if ((latest > m_L2FilterIndex) and h_L25TagInfo.isValid() and h_L25JetTags.isValid()) {
-    unsigned int size = std::min(h_L25TagInfo->size(), m_size);
+    unsigned int size = std::min((unsigned int) h_L25TagInfo->size(), m_size);
     for (unsigned int i = 0; i < size; ++i) {
       const reco::SoftLeptonTagInfo & info   = (*h_L25TagInfo)[i];
       const reco::Jet    & jet = * info.jet();
@@ -273,7 +273,7 @@ void HLTMonBTagMuSource::analyze(const edm::Event & event, const edm::EventSetup
     }
   }
   if ((latest > m_L25FilterIndex) and h_L3TagInfo.isValid() and h_L3JetTags.isValid()) {
-    unsigned int size = std::min(h_L3TagInfo->size(), m_size);
+    unsigned int size = std::min((unsigned int) h_L3TagInfo->size(), m_size);
     for (unsigned int i = 0; i < size; ++i) {
       const reco::SoftLeptonTagInfo & info   = (*h_L3TagInfo)[i];
       const reco::Jet    & jet = * info.jet();
